Add tweet prep helper method file
